Changes to Display and Data Functions
Condensed the information provided on the table of projects while leaving the "show" page intact with all the information.

Added thumbnail on the index for each project.

Fixed relation integrity issue with cascading deletion.

Slimmed down the index views by migrating code to the controller.

Added seeding to the Procfile.